Update my code to newest

+++ src/main/java/cn/nukkit/Server.java
@@ -1509,7 +1509,7 @@ public class Server {
     }
 
     public boolean generateLevel(String name) {
-        return this.generateLevel(name, (int) new java.util.Random().nextLong());
+        return this.generateLevel(name, new java.util.Random().nextLong());
     }
 
     public boolean generateLevel(String name, long seed) {
